SCRUM-110 | Litle Fix
